app/requirements.txt
discord.py[voice]==1.4.1
aioredis==1.3.1
marshmallow==3.7.1
